+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Azure Tools CRDs
 
 
+## [0.3.2] - 2020-04-28
+
+### Changed
+
+- Relax `AzureConfig` CRD validation.
+
+
+
 ## [0.3.1] - 2020-04-22
 
 ### Added
@@ -148,8 +156,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/apiextensions/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/giantswarm/apiextensions/compare/v0.3.2...HEAD
 
+[0.3.2]: https://github.com/giantswarm/apiextensions/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/giantswarm/apiextensions/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/giantswarm/apiextensions/compare/v0.2.6...v0.3.0
 [0.2.6]: https://github.com/giantswarm/apiextensions/compare/v0.2.5...v0.2.6